requirements.txt
requests
python-dotenv
PyMuPDF
openai
apscheduler
